requirements.txt
flask==2.3.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
schedule==1.2.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2 